Factor out utilities into their own files
