+++ packages/playground/src/utils/is-mac.ts
@@ -1,0 +1,3 @@
+export const isMac =
+  typeof navigator !== "undefined" &&
+  navigator.platform.toUpperCase().indexOf("MAC") >= 0;
