+++ packages/playground/src/utils/object.ts
@@ -1,0 +1,18 @@
+export function objectWithKey<T extends string>(
+  value: unknown,
+  key: T,
+): value is { [K in T]: unknown } {
+  return typeof value === "object" && value !== null && key in value;
+}
+
+export function objectWithKeyAndValue<T extends string, V>(
+  value: unknown,
+  key: T,
+  expectedValue: V,
+): value is { [K in T]: V } {
+  return objectWithKey(value, key) && value[key] === expectedValue;
+}
+
+export function objectHasStack(error: unknown): error is { stack: string } {
+  return objectWithKey(error, "stack") && typeof error.stack === "string";
+}
