+++ packages/playground/src/utils/url.ts
@@ -1,0 +1,7 @@
+export function getPathFromUrl(url: string): string | null {
+  try {
+    return new URL(url).pathname;
+  } catch (error) {
+    return null;
+  }
+}
